Split geometry into its own crate.
Layout and generating vector paths for a diagram were slightly misplaced in the core.
They should also not be in the webclient crate since the graphical aspects are independent
of the web user interface: We might want to generate SVGs or TikZ images from the command
line.

Started to use the `euclid` crate for geometry and `lyon` for vector paths. This enables us
to perform hit testing for interactive diagrams independently of browser APIs, so we are free
to merge graphical elements in the SVG for performance. Hit testing using `lyon` is a lot
faster than using the canvas and `Path2d` API from the browser, probably due to communication
overhead between JS and wasm. More performance probably can be had by computing and caching bounding
boxes, but it is already acceptable in debug builds and very good in release builds.

Wires in the SVG remain to be merged and since there are a lot of wire fragments in long identity
wires this would shrink the SVG significantly. There is no support for wire crossings yet, so
it might be best to delay this optimization until we figured out a good way to represent crossings.