Fixed and cleaned up the panzoom component.
It looks like mobile Safari does not support onpointermove at the moment, so I switched to separate touch and mouse events.

Using the `movement` fields instead of the `client` fields on the mouse move event led to the cursor and the element getting out of sync.

Smoke tested in Firefox on Linux and Safari on iOS.