Typechecking for homotopies and bug fixes.
Fixed bug in normalisation: The early return optimization that checks if
one arrow in the sink is an identity map returned only the rewrite part
of the arrow as a factor and ignored the degeneracy part. In the case
that this optimization applies, the degeneracy parts should be globular,
so we can fix this by converting the entire sink arrow to a rewrite.

Fixed bug in typechecking: the slices of `Embedding::Singular` start at
the height of the embedding, but this wasn't taken in account when
indexing the slice array.

Fixed broken pretty printing for `Diagram` and `Rewrite` caused by the
hash-consing type.