Fixed bug in type checking.
In the restriction of a rewrite the start height of the subdiagram that
we restrict to is subtracted from the index of the rewrite's cones.
Before this fix the height in the target diagram was subtracted, but the
cone's index is located in the source diagram. This is fixed by
translating the height back to the source diagram of the rewrite.