Fixes #41.
The optimization in normalization was invalid since it attempted to
convert a non-globular degeneracy to a rewrite. This occured since it
falsely identified one of the sink arrows to be an identity, which is
not possible at the moment since degeneracies do not keep track of what
happens on regular levels and hence if they are identities on regular
levels. If this turns out to be a performance concern, we could consider
what it would take to make this optimization work properly.