Hashconsing for Cones
Previously, we were not able to hashcons cones because many cones with
logically identical content appeared at different places in a diagram,
as indicated by `cone.index`. The sharable cone fields have been
factored out into a ConeInternal type, behind a hashcons smart pointer
as with the rest of the hashconsed types. This change is expected to be
breaking with respect to serialisation format.

Closes #63