Updated serialization format.
The hashes that identify diagrams and their parts are now based on the
data and not on the id which incidentally has been assigned by the
hash-consing implementation. Together with sorting this should guarantee
that the same diagram serializes to exactly the same string. The
integers in those structures that are addressed by their hash are made
to be u32 instead of usize so that their hash is constant among
architectures.

Changed to (gzipped) JSON from messagepack.  JSON is human-readable, so
debugging and exploring the file format is simpler.  This already
uncovered a bug that left half of the 128 bit keys zero, which would not
have been clear at all in messagepack.