Add `Yew` keys to signature items and folders (#660)
This creates a new field for folders `id`, which is used in the creation
of keys for folders. In order for this change to play well with the old
serialisation code, this field is not deserialised (on import) and
instead regenerated whilst importing. This is not ideal but will work
fine for the period while we get ready to replace the existing
serialisation code.