Substantial improvements to signature UI (#632)
* Add toggle for signature previews

* Add `single_preview` to `GeneratorInfo`
* Add toggle in generators for `single_preview`
* Deal with long generator names (but not for <input>) (#545)

* Reorder generator preferences

* Recolor shape picker based on generator colour

* Add basic indicators for invertibility/framed

* Migrate to switch style inputs over checkboxes

* Use better colours for icons and generator prefs

* Improve dynamic buttons for folders

* Fix folder buttons when editing

* Fix shape picker colours

* Trigger item view mode switch on finish hex input

* Replace white generator preset color with grey

* Cache previews to avoid frequent redraws

* Improve hovering stability substantially

* Adjust cursors for generator settings

* Make preview caching overly careful

Ideally, these caches could be implemented in such a way that they are
only recalculated when absolutely necessary. Unfortunately, this seems
non-trivial in current implementation (we would have to diff parts of
the signature tree).
This commit functions by diffing the entire signature for each signature
item `update()` (ie. transitions between viewing, hovering and edit).

* Recolour dimension text for clarity

* Recolour disabled generator preferences

* Add `Yew` keys to signature items and folders (#660)

This creates a new field for folders `id`, which is used in the creation
of keys for folders. In order for this change to play well with the old
serialisation code, this field is not deserialised (on import) and
instead regenerated whilst importing. This is not ideal but will work
fine for the period while we get ready to replace the existing
serialisation code.

* Add underline to text inputs

* Add global setting for toggling preview visibility

* Refine default behaviour

* Default to single previews for 1d and 2d